Oliverw

> You're right, I've actually glanced over it in the next commit. Keep up the good work ! I'd send a VTC donation , you have any address or just use the one that's in the code? :)

Thanks a lot. Greatly appreciated!

+++ src/Miningcore/Api/Responses/GetMinerStatsResponse.cs
@@ -24,7 +24,7 @@ namespace Miningcore.Api.Responses
 
     public class MinerStats
     {
-        public ulong PendingShares { get; set; }
+        public double PendingShares { get; set; }
         public decimal PendingBalance { get; set; }
         public decimal TotalPaid { get; set; }
         public decimal TodayPaid { get; set; }
