+++ src/Miningcore/Persistence/Model/Projections/MinerStats.cs
@@ -17,7 +17,7 @@ namespace Miningcore.Persistence.Model.Projections
 
     public class MinerStats
     {
-        public ulong PendingShares { get; set; }
+        public double PendingShares { get; set; }
         public decimal PendingBalance { get; set; }
         public decimal TotalPaid { get; set; }
         public decimal TodayPaid { get; set; }
